Add files via upload
Much more readability for variations chart with data labels which have bbox and same color as the lines. 
Used same ylim formatting for all charts.
Changed pastel to dark and dark to pastel between charts, in order to get the variations darker and easier to see with data labels.
Tiny improvements in writing.